Conflicts:
	fannie/modules/plugins2.0/HouseCoupon/HouseCouponEditor.php
	pos/is4c-nf/graphics/is4c.gif
	pos/is4c-nf/install/extra_config.php
	pos/is4c-nf/lib/Database.php
	pos/is4c-nf/lib/PrehLib.php
	pos/is4c-nf/lib/Scanning/SpecialUPCs/HouseCoupon.php

+++ fannie/admin/import/byLC.php
@@ -21,41 +21,43 @@
 
 *********************************************************************************/
 
-include('../../config.php');
-include($FANNIE_ROOT.'src/tmp_dir.php');
+include(dirname(__FILE__) . '/../../config.php');
+if (basename($_SERVER['PHP_SELF']) != basename(__FILE__)) {
+    return;
+}
 
 if (isset($_POST['MAX_FILE_SIZE'])){
-	$tpath = sys_get_temp_dir()."/vendorupload/";
-	if (!is_dir($tpath)) mkdir($tpath);
-	$dh = opendir($tpath);
-	while (($file = readdir($dh)) !== false) {
-		if (!is_dir($tpath.$file)) unlink($tpath.$file);
-	}
-	closedir($dh);
+    $tpath = sys_get_temp_dir()."/vendorupload/";
+    if (!is_dir($tpath)) mkdir($tpath);
+    $dh = opendir($tpath);
+    while (($file = readdir($dh)) !== false) {
+        if (!is_dir($tpath.$file)) unlink($tpath.$file);
+    }
+    closedir($dh);
 
-	$tmpfile = $_FILES['upload']['tmp_name'];
-	$path_parts = pathinfo($_FILES['upload']['name']);
-	if ($path_parts['extension'] == "zip"){
-		move_uploaded_file($tmpfile,$tpath."CAP.zip");
-		$output = system("unzip {$tpath}CAP.zip -d $tpath &> /dev/null");
-		unlink($tpath."CAP.zip");
-		$dh = opendir($tpath);
-		while (($file = readdir($dh)) !== false) {
-			if (!is_dir($tpath.$file)) rename($tpath.$file,$tpath."lcimp.csv");
-		}
-		closedir($dh);
-	}
-	else {
-		move_uploaded_file($tmpfile, $tpath."lcimp.csv");
-	}
-	header("Location: load.php");
+    $tmpfile = $_FILES['upload']['tmp_name'];
+    $path_parts = pathinfo($_FILES['upload']['name']);
+    if ($path_parts['extension'] == "zip"){
+        move_uploaded_file($tmpfile,$tpath."CAP.zip");
+        $output = system("unzip {$tpath}CAP.zip -d $tpath &> /dev/null");
+        unlink($tpath."CAP.zip");
+        $dh = opendir($tpath);
+        while (($file = readdir($dh)) !== false) {
+            if (!is_dir($tpath.$file)) rename($tpath.$file,$tpath."lcimp.csv");
+        }
+        closedir($dh);
+    }
+    else {
+        move_uploaded_file($tmpfile, $tpath."lcimp.csv");
+    }
+    header("Location: load.php");
 }
 else {
 
-	/* html header, including navbar */
-	$page_title = "Fannie - Import info";
-	$header = "Import CSV info";
-	include($FANNIE_ROOT."src/header.html");
+    /* html header, including navbar */
+    $page_title = "Fannie - Import info";
+    $header = "Import CSV info";
+    include($FANNIE_ROOT."src/header.html");
 ?>
 <form enctype="multipart/form-data" action="byLC.php" method="post">
 <input type="hidden" name="MAX_FILE_SIZE" value="2097152" />
@@ -63,7 +65,7 @@ Filename: <input type="file" id="file" name="upload" />
 <input type="submit" value="Upload File" />
 </form>
 <?php
-	/* html footer */
-	include($FANNIE_ROOT."src/footer.html");
+    /* html footer */
+    include($FANNIE_ROOT."src/footer.html");
 }
 ?>
