+++ fannie/cron/manage.cache.php
@@ -23,6 +23,9 @@
 
 /* HELP
 
+   @deprecated. ReportDataCache task handles
+   all cached data.
+
    manage.cache.php
 
    Delete daily & monthly cache files as needed
@@ -38,22 +41,30 @@ set_time_limit(0);
 
 $path = 'cache/cachefiles/daily/';
 $dh = opendir($FANNIE_ROOT.$path);
+if (!$dh) {
+    echo cron_msg("$FANNIE_ROOT.$path does not exist.");
+    exit;
+}
 while ( ($file = readdir($dh)) !== False){
-	if (is_file($FANNIE_ROOT.$path.$file))
-		unlink($FANNIE_ROOT.$path.$file);
+    if (is_file($FANNIE_ROOT.$path.$file))
+        unlink($FANNIE_ROOT.$path.$file);
 }
 closedir($dh);
 echo cron_msg("Cleared daily cache");
 
 if (date('j') == 1){
-	$path = 'cache/cachefiles/monthly/';
-	$dh = opendir($FANNIE_ROOT.$path);
-	while ( ($file = readdir($dh)) !== False){
-		if (is_file($FANNIE_ROOT.$path.$file))
-			unlink($FANNIE_ROOT.$path.$file);
-	}
-	closedir($dh);
-	echo cron_msg("Cleared monthly cache");
+    $path = 'cache/cachefiles/monthly/';
+    $dh = opendir($FANNIE_ROOT.$path);
+    if (!$dh) {
+        echo cron_msg("$FANNIE_ROOT.$path does not exist.");
+        exit;
+    }
+    while ( ($file = readdir($dh)) !== False){
+        if (is_file($FANNIE_ROOT.$path.$file))
+            unlink($FANNIE_ROOT.$path.$file);
+    }
+    closedir($dh);
+    echo cron_msg("Cleared monthly cache");
 }
 
 ?>
